Unifies file descriptor types to isize. Additionally, refines lifetime annotations for create_lazy_scope to improve type safety with generic closures and adds the Win32_Security feature to the windows-sys dependency. Implements ElfObject and Drop for ElfFile on windows.

+++ src/mmap.rs
@@ -64,7 +64,7 @@ pub trait Mmap {
         prot: ProtFlags,
         flags: MapFlags,
         offset: usize,
-        fd: Option<i32>,
+        fd: Option<isize>,
         need_copy: &mut bool,
     ) -> Result<NonNull<c_void>>;
